Create README.md
SHA Food Delivery: Helping Everyone Eat Well
SHA Food Delivery is like a food delivery app, but we're also trying to make the world a better place! We want to help everyone have enough food to eat, and that's why we're connected to the "Zero Hunger" goal.
 
Here's how we do it:
 
- Less Food Waste: We help restaurants sell food that might go bad, so it doesn't get thrown away.
- Supporting Local Farms: We work with farmers near you, so you get fresh food and they get more business.
- Healthy Choices: We offer lots of yummy and healthy food, so you can eat well and feel good.
- Easy Food For Everyone: We make it easy for everyone to get food, even if they don't have a car or a lot of time.
- Helping Small Businesses: We help small restaurants grow their business, which means more jobs for people.
 
We also try to be good to the planet:
 
- Less Pollution: We try to deliver food in a way that doesn't pollute too much.
- Good Ingredients: We work with companies that care about the environment.
- Helping People in Need: We work with groups that help people who don't have enough food.
 
So, when you order from SHA Food Delivery, you're not just getting a delicious meal, you're helping us make the world a better pla
Let's make sure everyone has enough to eat!